Don't create files when they already exist

+++ index.js
@@ -8,6 +8,8 @@ if(args[0] === "create-controller" || args[0]=== "controller-generator"){
 }
 
 const pathExists = `./Controllers`;
+let filesCreated = 0;
+let rejectedFiles = 0;
 
   try {
     if (!fs.existsSync(pathExists)) {
@@ -18,20 +20,39 @@ const pathExists = `./Controllers`;
   }
 
   args.map(function(res) {
-    fs.writeFile(`./Controllers/${res}Controller.js`, "", function(
-      err
-    ) {
-      if (err) throw err;
-    });
-
-    fs.copyFile(
-      `${__dirname}/controllerHolder.js`,
-      `./Controllers/${res}Controller.js`,
-
-      function(err) {
+    // check if the file name does not already exist
+    if(fs.existsSync(`./Controllers/${res}Controller.js`)) {
+      rejectedFiles++;
+    } else {
+      fs.writeFile(`./Controllers/${res}Controller.js`, "", function(
+        err
+      ) {
         if (err) throw err;
-      }
-    );
+      });
+  
+      fs.copyFile(
+        `${__dirname}/controllerHolder.js`,
+        `./Controllers/${res}Controller.js`,
+  
+        function(err) {
+          if (err) throw err;
+        }
+      );
+      filesCreated++;
+    }
   });
-  const fileCount = args.length > 1 ? "Files" : "File";
-  console.log(`${args.length} ${fileCount} created successfully.`);
+
+  // show the user the number of files created
+  if(rejectedFiles > 0) {
+    rejectedFiles === 1 ? console.log(`${rejectedFiles} FILE NOT CREATED : The file you are trying to create already exists`) 
+      : console.log(`${rejectedFiles} FILES NOT CREATED : The files you are trying to create already exist`);
+    
+  }
+
+  if(filesCreated > 0) {
+    if(filesCreated > 0) {
+      filesCreated === 1 ?  console.log(`${filesCreated} file created successfully.`) 
+        : console.log(`${filesCreated} files created successfully.`);
+    }
+  }
+
